requirements-v5.1.txt
# ============================================================================
# Golden Architecture V5.1 - Production Dependencies
# ============================================================================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis (for idempotency)
redis[hiredis]==5.0.1

# NATS Messaging
nats-py==2.6.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# JSON Schema Validation
jsonschema==4.20.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Metrics & Monitoring
prometheus-client==0.19.0

# Circuit Breaker
tenacity==8.2.3  # Optional: alternative to custom implementation

# Container Management (for sandbox)
docker==7.0.0  # Optional: if using Docker SDK

# Data Processing
pandas==2.1.4  # For analytics
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing

# Type Checking (Development)
mypy==1.7.1
types-redis==4.6.0.11

# Code Quality (Development)
black==23.12.0
ruff==0.1.8
isort==5.13.2

# ============================================================================
# Optional: Kubernetes Client (if running in K8s)
# ============================================================================
# kubernetes==28.1.0

# ============================================================================
# Optional: Distributed Tracing
# ============================================================================
# opentelemetry-api==1.21.0
# opentelemetry-sdk==1.21.0
# opentelemetry-instrumentation-fastapi==0.42b0

# ============================================================================
# Optional: Advanced Security
# ============================================================================
# cryptography==41.0.7
# pyjwt==2.8.0

# ============================================================================
# Version Notes
# ============================================================================
# - All versions pinned for reproducible builds
# - Security: Latest stable versions as of 2025-01
# - Python 3.11+ required for best performance
